+++ src/geometry/mesh.rs
@@ -6,7 +6,6 @@ use ultraviolet::Vec3;
 use crate::geometry::{Boxable, Intersectable, Intersection};
 use crate::geometry::aabb::Aabb;
 use crate::geometry::ray::Ray;
-use crate::Float;
 
 #[derive(Debug, Deserialize, Serialize)]
 pub enum DrawMode {
@@ -64,9 +63,9 @@ impl Mesh {
             );
 
             // scatter face normals to vertex normals
-            self.vertices.index_mut(triangle.index0).normal += w0 as f32 * triangle.normal;
-            self.vertices.index_mut(triangle.index1).normal += w1 as f32 * triangle.normal;
-            self.vertices.index_mut(triangle.index2).normal += w2 as f32 * triangle.normal;
+            self.vertices.index_mut(triangle.index0).normal += w0 * triangle.normal;
+            self.vertices.index_mut(triangle.index1).normal += w1 * triangle.normal;
+            self.vertices.index_mut(triangle.index2).normal += w2 * triangle.normal;
         }
 
 
@@ -77,8 +76,8 @@ impl Mesh {
     }
 
     pub fn compute_bounding_box(&mut self) {
-        let mut min = Vec3::one() * Float::INFINITY as f32;
-        let mut max = Vec3::one() * Float::NEG_INFINITY as f32;
+        let mut min = Vec3::one() * f32::INFINITY;
+        let mut max = Vec3::one() * f32::NEG_INFINITY;
 
         for vertex in &self.vertices {
             min = min.min_by_component(vertex.position);
@@ -99,14 +98,14 @@ impl Mesh {
     ///
     /// # Returns
     /// weights to be used for vertices 0, 1 and 2
-    fn angle_weights(p0: Vec3, p1: Vec3, p2: Vec3) -> (Float, Float, Float) {
+    fn angle_weights(p0: Vec3, p1: Vec3, p2: Vec3) -> (f32, f32, f32) {
         let e01 = (p1 - p0).normalized();
         let e12 = (p2 - p1).normalized();
         let e20 = (p0 - p2).normalized();
 
-        let w0 = Float::max(-1.0, Float::min(1.0, e01.dot(-e20) as Float)).acos();
-        let w1 = Float::max(-1.0, Float::min(1.0, e12.dot(-e01) as Float)).acos();
-        let w2 = Float::max(-1.0, Float::min(1.0, e20.dot(-e12) as Float)).acos();
+        let w0 = f32::max(-1.0, f32::min(1.0, e01.dot(-e20))).acos();
+        let w1 = f32::max(-1.0, f32::min(1.0, e12.dot(-e01))).acos();
+        let w2 = f32::max(-1.0, f32::min(1.0, e20.dot(-e12))).acos();
 
         (w0, w1, w2)
     }
