+++ src/geometry/plane.rs
@@ -3,19 +3,18 @@ use ultraviolet::Vec3;
 
 use crate::geometry::{Aabb, AngularExt, Boxable, Intersectable, Intersection};
 use crate::geometry::ray::Ray;
-use crate::Float;
 
 /// A geometrical plane.
 #[derive(Debug, Default, Deserialize, Serialize)]
 pub struct Plane {
     /// The distance of the plane into the normal direction.
-    pub d: Float,
+    pub d: f32,
     /// The normal of a plane.
     pub normal: Vec3,
 }
 
 impl Plane {
-    pub fn new(d: Float, normal: Vec3) -> Self {
+    pub fn new(d: f32, normal: Vec3) -> Self {
         Self {
             d,
             normal: normal.normalized(),
@@ -26,17 +25,17 @@ impl Plane {
     #[inline]
     #[must_use]
     pub fn project(&self, point: Vec3) -> Vec3 {
-        point - (self.normal * self.distance(point) as f32)
+        point - (self.normal * self.distance(point))
     }
 
     /// Calculates the distance of the given point to the plane.
     #[inline]
     #[must_use]
-    pub fn distance(&self, point: Vec3) -> Float {
-        let v = point - (self.normal * self.d as f32);
+    pub fn distance(&self, point: Vec3) -> f32 {
+        let v = point - (self.normal * self.d);
         let angle = self.normal.angle_to(v);
 
-        angle.cos() * v.mag() as Float
+        angle.cos() * v.mag()
     }
 }
 
@@ -46,7 +45,7 @@ impl From<(Vec3, Vec3)> for Plane {
         let normal = pos_normal.1.normalized();
 
         let angle = position.angle_to(normal);
-        let d = angle.cos() * position.mag() as Float;
+        let d = angle.cos() * position.mag();
 
         Self { d, normal }
     }
@@ -60,13 +59,13 @@ impl Boxable for Plane {
 
 impl Intersectable<Ray> for Plane {
     fn intersects(&self, ray: &Ray) -> Option<Intersection> {
-        let denom = self.normal.dot(ray.direction) as Float;
-        if denom.abs() <= Float::EPSILON {
+        let denom = self.normal.dot(ray.direction);
+        if denom.abs() <= f32::EPSILON {
             return None;
         }
 
-        let t = -(self.normal.dot(ray.origin) as Float + self.d) / denom;
-        if t <= Float::EPSILON {
+        let t = -(self.normal.dot(ray.origin) + self.d) / denom;
+        if t <= f32::EPSILON {
             return None;
         }
