+++ src/geometry/sphere.rs
@@ -3,7 +3,6 @@ use ultraviolet::Vec3;
 
 use crate::geometry::{Aabb, Boxable, Intersectable, Intersection};
 use crate::geometry::ray::Ray;
-use crate::Float;
 
 /// A geometrical sphere.
 #[derive(Debug, Default, Deserialize, Serialize)]
@@ -11,18 +10,18 @@ pub struct Sphere {
     /// The center of the sphere.
     pub center: Vec3,
     /// The radius of the sphere.
-    pub radius: Float,
+    pub radius: f32,
 }
 
 impl Sphere {
-    pub fn new(center: Vec3, radius: Float) -> Self {
+    pub fn new(center: Vec3, radius: f32) -> Self {
         Sphere { center, radius }
     }
 }
 
 impl Boxable for Sphere {
     fn bounding_box(&self) -> Option<Aabb> {
-        let offset = Vec3::one() * self.radius as f32;
+        let offset = Vec3::one() * self.radius;
 
         Some(Aabb::new(self.center - offset, self.center + offset))
     }
@@ -32,9 +31,9 @@ impl Intersectable<Ray> for Sphere {
     fn intersects(&self, ray: &Ray) -> Option<Intersection> {
         let oc = ray.origin - self.center;
 
-        let a = ray.direction.mag_sq() as Float;
-        let b = 2.0 * oc.dot(ray.direction) as Float;
-        let c = oc.mag_sq() as Float - self.radius * self.radius;
+        let a = ray.direction.mag_sq();
+        let b = 2.0 * oc.dot(ray.direction);
+        let c = oc.mag_sq() - self.radius * self.radius;
         let discriminant = b * b - 4.0 * a * c;
 
         if discriminant <= 0.0 {
